requirements.txt
Flask==2.3.2
mysql-connector-python
Flask-Mail
